data validation tamamlandı

+++ api/Controller/CommentController.cs
@@ -26,14 +26,20 @@ namespace api.Controller
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var comments = await _commentRepo.GetAllAsync();     
             var commentDto = comments.Select(s => s.ToCommentDto());
             return Ok(commentDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var comment = await _commentRepo.GetByIdAsync(id);
             if(comment == null){
                 return NotFound();
@@ -41,9 +47,13 @@ namespace api.Controller
             return Ok(comment.ToCommentDto());
         }
         
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
         {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             if(!await _stockRepo.StockExist(stockId)){
                 return BadRequest("Hisse Bulunamadı!");
             }
@@ -53,9 +63,13 @@ namespace api.Controller
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate());
 
             if(comment == null){
@@ -65,9 +79,13 @@ namespace api.Controller
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             var commentModel = await _commentRepo.DeleteAsync(id);
             if(commentModel == null){
                 return NotFound("Yorum Bulunamadı!");
